Validation so that users cannot change, delete and mark tasks of other users as complete, changing priority return and Task status

+++ src/main/java/com/todoapi/services/TaskService.java
@@ -5,15 +5,33 @@ import java.util.Optional;
 
 import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 
 import com.todoapi.entities.Task;
+import com.todoapi.entities.User;
 import com.todoapi.repositories.TaskRepository;
+import com.todoapi.repositories.UserRepository;
+import com.todoapi.security.JWTAuthenticationFilter;
+import com.todoapi.security.JWTUtil;
+import com.todoapi.security.UserSS;
+import com.todoapi.services.authentication.UserServiceAuthentication;
 
 @Service
 public class TaskService {
 	@Autowired
 	private TaskRepository repository;
+	
+	@Autowired
+	private UserService userService;
+	
+	@Autowired
+	private JWTUtil jwtUtil;
+	
+	@Autowired
+	private UserRepository userRepository;
 	
 	public List<Task> findAll(){
 		return repository.findAll();
@@ -26,18 +44,25 @@ public class TaskService {
 	
 	public Task insert(Task obj) {
 		obj.setId(null);
+		obj.setUser(userService.findByEmail(jwtUtil.getUsername(JWTAuthenticationFilter.token)));
 		obj = repository.save(obj);
 		return obj;
 	}
 	
-		public Task update(Task obj) {
-		Task newObj = findById(obj.getId());
-		updateData(newObj, obj);
-		return repository.save(newObj);	
+	public Task update(Task obj) {
+		if(validateResponsibleForTheTask(jwtUtil.getUsername(JWTAuthenticationFilter.token)) == null) {
+			throw new RuntimeException("Could not execute command, unauthorized user");
+		}
+			Task newObj = findById(obj.getId());
+			updateData(newObj, obj);
+			return repository.save(newObj);	
 		
 	}
 
-	public void delete(long id) {
+	public void delete(long id, Task task) {
+		if(validateResponsibleForTheTask(jwtUtil.getUsername(JWTAuthenticationFilter.token)) == null) {
+			throw new RuntimeException("Could not execute command, unauthorized user");
+		}
 		repository.deleteById(id);
 	}
 	
@@ -47,4 +72,36 @@ public class TaskService {
 		newObj.setStatus(obj.getStatus());
 	}
 
+	public Task updateTaskFinish(Task obj) {
+		if(validateResponsibleForTheTask(jwtUtil.getUsername(JWTAuthenticationFilter.token)) == null) {
+			throw new RuntimeException("Could not execute command, unauthorized user");
+		}
+			Task newObj = findById(obj.getId());
+			updateTaskState(newObj, obj);
+			return repository.save(newObj);
+	}
+	
+	private void updateTaskState(Task newObj, Task obj) {
+		newObj.setStatus(obj.getStatus());
+	}
+	
+	public Page<Task> listTasks(Integer pages, Integer linesPerPage, String orderBy, String direction){
+		UserSS userSS = UserServiceAuthentication.authenticated();
+		if(userSS == null) {
+			throw new RuntimeException("unauthorized user");
+		}
+		
+		User user = userRepository.findById(userSS.getId()).get();
+		
+		PageRequest pageRequest = PageRequest.of(pages, linesPerPage, Direction.valueOf(direction), orderBy);
+		
+		return repository.findByUser(user, pageRequest);
+	}
+	
+	public Boolean validateResponsibleForTheTask(String email) {
+		if(email.equals(jwtUtil.getUsername(JWTAuthenticationFilter.token))) {
+			return true;
+		}
+		return null;
+	}
 }
